Create higgsbosonplotting.py
This is the initial code for building the plots being used. It was made from this example showing how to do so using CMS data from 2012 involving dimuon particle decays. It can be accessed here: 
https://github.com/cms-opendata-education/cms-jupyter-materials-english/blob/master/Exercises-with-open-data/Advanced/Normfit-transversemomentum%2Bpseudorapidity.ipynb

There may be changes made depending on what method is to be used in making these graphs.